+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 -   Correct a warning about fs module for front-end builds.
 -   Remove a circle dependency (dataframe.js -> groupedDataframe.js -> dataframe.js)
 -   Add .tail, .head, .slice, .getRow, .setRow methods.
+-   A filtered DataFrame with empty rows will now returns its columns.
 
 ---